split cut/mod.rs into submodules
Break the ~1500-line cut/mod.rs into face, geometry, classify, snap,
tables, and assemble submodules, moving each test alongside the code
it exercises.
